recruiter-pro/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1
anthropic==0.26.0
python-multipart==0.0.9
python-docx==1.1.2
